Service depends on the task execution policy
Fixes this error message from the ECS console under Events:

service X failed to launch a task with (error ECS was unable to assume the role 'arn:aws:iam::<account_id>:role/X-task-execution-role' that was provided for this task. Please verify that the role being passed has the proper trust relationship and permissions and that your IAM user has permissions to pass this role.).