Remove target group name
The target group name was added for readability but was orignially
excluded because of problems destroying a target group while a listener
was attached. See comment in code. This was removed again in case this
was still an issue and becuase of a length limit of 32 chars for the
name parameter which was causing problems